+++ src/main/java/generator/LineGenerator.java
@@ -1,4 +1,4 @@
-package util;
+package generator;
 
 public interface LineGenerator {
