+++ src/test/java/domain/PlayersMakerTest.java
@@ -1,9 +1,10 @@
-package util;
+package domain;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
 import domain.Player;
 import domain.Players;
+import domain.PlayersMaker;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
