+++ src/test/java/domain/PlayersTest.java
@@ -10,7 +10,6 @@ import java.util.List;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import util.PlayersMaker;
 
 public class PlayersTest {
